+++ src/main/java/engine/model/physicalEngine/shape/Shape.java
@@ -1,8 +1,8 @@
-package engine.model.physicalEngine.entity.shape;
+package engine.model.physicalEngine.shape;
 
 import java.util.List;
 
-import engine.model.physicalEngine.entity.temp.Position;
+import engine.model.physicalEngine.shape.temp.Position;
 
 public interface Shape {
     void setCenter(Position point);
