+++ src/test/java/engine/physicalEngine/entity/shape/CircleTest.java
@@ -1,0 +1,23 @@
+package engine.physicalEngine.entity.shape;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import org.junit.jupiter.api.Test;
+
+import engine.model.physicalEngine.shape.Circle;
+
+public class CircleTest {
+
+    @Test
+    void testGetApex() {
+        float posX = 10;
+        float posY = 10;
+        // Circle circle = new Circle(posX, posY, 5);
+        // assertEquals(circle.getApex().isEmpty(), true);
+        // assertEquals(circle.getCenter().getX(), 0.0);
+        // assertEquals(circle.getCenter().getY(), 0.0);
+    }
+
+    @Test
+    void testgetCenter(){}
+}
